Revert "filtros ahora funcionan"
