--- client/src/components/filters/filters.module.css
@@ -1,40 +1,0 @@
-.Container_filter{
-	display: flex;
-	justify-content: center;
-	flex-direction: column;
-	align-items: center;
-}
-
-
-
-
-
-.CategoryZone{
-    background-color: #fff;
-    border: 1px solid #8f8f8f;
-    font-family: Helvetica;
-    font-size: 20px;
-    margin: 0 10px;
-    padding: 8px 25px;
-}
-
-.OrderZone{
-	margin: 30px 0;
-}
-
-.OrderSelect{
-    background-color: #fff;
-    border: 1px solid #8f8f8f;
-    font-family: Helvetica;
-    font-size: 15px;
-    margin: 0 10px;
-    padding: 4px 15px;		
-}
-
-
-
-
-
-
-
-
