+++ client/src/containers/feedpage/FeedPage.jsx
@@ -3,26 +3,45 @@ import Card from "../../components/cards/card";
 import { useDispatch, useSelector } from "react-redux";
 import s from "./feedpage.module.css";
 import {
+  GetAllPosts,
+  GetAllCategories,
   setPage,
+  priceOrder,
   CleanPosts,
+  antOrder,
+  likesOrder,
   resetPage,
+  Filter,
 } from "../../redux/actions";
 import Categories from "../../components/categories/categories";
 import InfiniteScroll from "react-infinite-scroll-component";
+import { useState } from "react";
 import { useEffect } from "react";
+import { useLocation } from "react-router-dom";
 import NotFound from "../../components/notFound/NotFound"
-import Filters from "../../components/filters/filters"
 
 
 const PrincipalPage = () => {
   const dispatch = useDispatch();
   const allPosts = useSelector((state) => state.posts);
-  const hasMore = useSelector((state) => state.hasMore);
+  const allCategories = useSelector((state) => state.categories);
+  const length = useSelector((state) => state.length);
+  const [hasMore, setHasMore] = useState(true);
+  const page = useSelector((state) => state.page);
+  const name = useLocation();
+  
   const filterState = useSelector((state) => state.filter);
+  // const [filterStateLocal, setFilterStateLocal] = useState(false);
+
+  const [filterName, setFilterName] = useState('');
+
+  const [order, setOrder] = useState('');
+  const [refresh, setRefresh] = useState('');
   const foundOrNot = useSelector((state) => state.loader);
 
   useEffect(() => {
     return () => {
+      setHasMore(true);
       dispatch(resetPage());
       dispatch(CleanPosts());
     };
@@ -35,9 +54,104 @@ const PrincipalPage = () => {
   }, [filterState]);  
 
 
+  useEffect(() => {
+ 
+
+    // console.log(filterStateLocal)
+    // console.log('filtro global'+ filterState)
+
+    if(filterState){
+
+      if (filterName == 'price') {
+        dispatch(priceOrder(order, page));
+      }
+
+      if (filterName == 'antiguedad') {
+        dispatch(antOrder(order, page)); 
+      }
+      if (filterName == 'likes') {
+        dispatch(likesOrder(order, page)); 
+      }
+    }
+
+    if (!filterState) {
+      dispatch(GetAllPosts(page, name.search));
+    }
+
+    
+      dispatch(GetAllCategories());
+      if (page !== 0) {
+        if (page === Math.floor(length / 12)) setHasMore(false);          
+      }
+
+
+
+  }, [dispatch, page, name.search, length, order, refresh, foundOrNot,filterState]);
+
+
+
+
+  let orderByPrice=(e) =>{
+    e.preventDefault();
+    dispatch(resetPage())
+    dispatch(Filter())
+    setFilterName('price')
+    setOrder(e.target.value)
+    setRefresh('pricee')
+     
+  }
+
+
+
+  function orderByAnt(e) {
+    e.preventDefault();
+    dispatch(resetPage())
+    dispatch(Filter())
+    setFilterName('antiguedad')
+    setOrder(e.target.value)
+    setRefresh('ant') 
+       
+  }
+
+  function orderByLikes(e) {
+    e.preventDefault();
+    dispatch(resetPage())
+    dispatch(Filter())
+    setFilterName('likes')
+    setOrder(e.target.value)  
+    setRefresh('like')
+    
+  }
+
 
   return (
     <div>
+      <select onChange={(e) => orderByPrice(e)}>
+        <option selected disabled>
+          ORDENAR POR PRECIO
+        </option>
+        <option value="DESC">Mayor a menor</option>
+        <option value="ASC">Menor a mayor</option>
+      </select>
+
+
+      <select onChange={(e) => orderByAnt(e)}>
+        <option selected disabled>
+          ORDENAR POR ANTIGUEDAD
+        </option>
+        <option value="DESC" >Más recientes</option>
+        <option value="ASC">Más antiguos</option>
+      </select>
+
+      <select onChange={(e) => orderByLikes(e)}>
+        <option selected disabled>
+          ORDENAR POR LIKES
+        </option>
+        <option value="ASC">Mejores valorados</option>
+        <option value="DESC">Peores valorados</option>
+      </select>
+
+
       <InfiniteScroll
         dataLength={allPosts.length}
         hasMore={hasMore}
@@ -50,10 +164,11 @@ const PrincipalPage = () => {
         }
       >
         <div className={s.FeedPage}>
-
-          <Filters />
-
-
+          <div className={s.CategoryZone}>
+            {allCategories?.map((cat) => (
+              <Categories title={cat.title} />
+            ))}
+          </div>
       {foundOrNot?
         <div><h2>Cargando...</h2></div>
         :allPosts.length?
