+++ client/src/redux/actions/index.js
@@ -208,14 +208,6 @@ export const CleanPosts = () => {
   };
 };
 
-export const PageNumber = () => {
-  return {
-    type: "PageNumber",
-  };
-};
-
-
-
 export const categoryFilter = (order, page = 0) => {
   return async function (dispatch) {
     let filterCategory = await axios.get(
