+++ client/src/redux/reducer/index.js
@@ -11,9 +11,8 @@ const initialState = {
   search: '',
   compressedPost: '',
   filter: false,
-  loader:true,
-  hasMore:true,
-  pageNumber:[]
+  countries:[],
+  loader:true
 };
 
 
@@ -25,16 +24,15 @@ function rootReducer(state = initialState, action) {
         return {
           ...state,
           posts: action.artWorks,
-          length: action.length,
+          length: action.counter,
           filter: false,
           loader: false,
-          hasMore:true
         };
       } else {
         return {
           ...state,
           posts: [...state.posts, ...action.artWorks],
-          length: action.length,
+          length: action.counter,
           filter: false,
           loader:false
         };
@@ -48,7 +46,6 @@ function rootReducer(state = initialState, action) {
           length: action.counter,
           filter: false,
           loader: false,
-          hasMore:true
         };
       } else {
         return {
@@ -84,31 +81,12 @@ function rootReducer(state = initialState, action) {
         detail: action.payload,
         filter:false
       };
-      
-    case "PageNumber":
-      const pageNumbers = []
-      for (let i = 1; i<= Math.ceil(state.length/12); i++){
-          pageNumbers.push(i)
-      }    
-        return {
-          ...state,
-          pageNumber: pageNumbers
-        };    
-
+    
     case "setPage":
-    if (state.page === Math.floor(state.length / 12)) {
       return {
         ...state,
         page: ++state.page,
-        hasMore:false
       };
-    } else {
-      return {
-        ...state,
-        page: ++state.page,
-        hasMore:true
-      };      
-    }
     
     case "resetPage":
       return {
@@ -154,8 +132,7 @@ function rootReducer(state = initialState, action) {
           posts: action.payload.Artworks,
           length: action.payload.counter,
           filter: true,
-          loader:false,
-          hasMore:true
+          loader:false
         };
       } else {
           return {
@@ -167,7 +144,6 @@ function rootReducer(state = initialState, action) {
           };              
       }
 
-    
     
     case "Filter":
       return {
@@ -181,7 +157,6 @@ function rootReducer(state = initialState, action) {
         filter: false,
       };
     
-    
     case "AntOrder":
       if (state.page === 0) {
         return {
@@ -189,8 +164,7 @@ function rootReducer(state = initialState, action) {
           posts: action.payload.Artoworks,
           length: action.payload.counter,
           filter: true,
-          loader:false,
-          hasMore:true
+          loader:false
         };
       } else {
         return {
@@ -211,8 +185,7 @@ function rootReducer(state = initialState, action) {
           posts: action.payload.Artworks,
           length: action.payload.counter,
           filter: true,
-          loader:false,
-          hasMore:true
+          loader:false
         };
       } else {
         return {
@@ -223,6 +196,25 @@ function rootReducer(state = initialState, action) {
           loader:false
         };
       };
+
+    case "CountryFilter":
+      if (state.page === 0) {
+        return {
+          ...state,
+          posts: action.payload,
+          length: action.payload.counter,
+          filter: true,
+          loader:false
+        };
+      } else {
+        return{
+          ...state,
+          posts: [...state.posts, ...action.payload],
+          length: action.payload.counter,
+          filter: true,
+          loader:false            
+        }
+      }
     
     case "CategoryFilter":
       if (state.page === 0) {
@@ -231,8 +223,7 @@ function rootReducer(state = initialState, action) {
           posts: action.payload.Artworks,
           length: action.payload.counter,
           filter: true,
-          loader:false,
-          hasMore:true
+          loader:false
         }
       } else {
         return{
@@ -244,7 +235,18 @@ function rootReducer(state = initialState, action) {
         }
       }
     
-   
+    case "Countries":
+      let allCountries = [];
+      let country = action.payload.map((e) => {
+        if (!allCountries.includes(e.country)) {
+          allCountries.push(e.country);
+        };
+      });
+      return {
+        ...state,
+        countries: allCountries,
+      };
+    
     case "REGISTER":
       return {...state};
